finance-cron/requirements.txt
requests>=2.31.0
PyYAML
APScheduler>=3.11.0
httpx>=0.28.0
PyJWT>=2.10.0
python-dotenv>=1.0.1
